BufferGroup.numberOfSamples is now a double (bug #343)

+++ core/film.h
@@ -132,7 +132,7 @@ protected:
     Scene *scene;
 
 	// Dade - 1.0 / (xResolution * yResolution)
-	float invSamplePerPass;
+	double invSamplePerPass;
 };
 
 // Image Pipeline Declarations
